refactor(gen-middleware): dedupe IP helpers, drop dead BYOP re-fetch, dedupe meter-slug mapping
- Move hashIp/stripIPv4MappedPrefix/expandIPv6/truncateIpToSubnet/bytesToHex into shared/client-ip.ts; track.ts and realtime.ts import them, deleting both local copies
- Drop resolveByopClientTracking redundant D1 re-fetch; auth-context byopClientName/byopClientUserId/byopClientKeyId already populate userTracking
- Add payerBucketToMeter helper to shared/billing/balance.ts; dedupe the local:bucket / v1:meter:slug ternary across track.ts, realtime.ts, and createBalanceCheckResult
- Output strings unchanged; no test edits
