+++ gen.pollinations.ai/src/middleware/track.ts
@@ -1,14 +1,17 @@
 import { getLogger } from "@logtape/logtape";
 import { AUTO_TOP_UP_THRESHOLD_POLLEN } from "@shared/billing/auto-top-up.ts";
+import { payerBucketToMeter } from "@shared/billing/balance.ts";
 import {
     handleBalanceDeduction,
     type MarkupResolution,
 } from "@shared/billing/track-helpers.ts";
-import { getRealClientIp } from "@shared/client-ip.ts";
 import {
-    apikey as apikeyTable,
-    user as userTable,
-} from "@shared/db/better-auth.ts";
+    getRealClientIp,
+    hashIp,
+    stripIPv4MappedPrefix,
+    truncateIpToSubnet,
+} from "@shared/client-ip.ts";
+import { user as userTable } from "@shared/db/better-auth.ts";
 import { PUBLIC_URLS } from "@shared/public-urls.ts";
 import type { Usage } from "@shared/registry/registry.ts";
 import {
@@ -189,10 +192,6 @@ export const track = (eventType: EventType) =>
                 } satisfies BalanceData;
 
                 const ipHash = await hashIp(clientIp, c.env.BETTER_AUTH_SECRET);
-                const byopClientTracking = await resolveByopClientTracking(
-                    db,
-                    byopClientKeyId,
-                );
 
                 // Deduct payer + credit dev before emitting the event so billing
                 // telemetry reflects the committed ledger state.
@@ -245,11 +244,7 @@ export const track = (eventType: EventType) =>
                 const committedBalanceTracking = payerBucket
                     ? {
                           ...balanceTracking,
-                          selectedMeterId: `local:${payerBucket}`,
-                          selectedMeterSlug:
-                              payerBucket === "tier"
-                                  ? "v1:meter:tier"
-                                  : "v1:meter:pack",
+                          ...payerBucketToMeter(payerBucket),
                       }
                     : balanceTracking;
 
@@ -263,10 +258,7 @@ export const track = (eventType: EventType) =>
                     eventType,
                     ipSubnet,
                     ipHash,
-                    userTracking: {
-                        ...userTracking,
-                        ...byopClientTracking,
-                    },
+                    userTracking,
                     balanceTracking: committedBalanceTracking,
                     requestTracking,
                     responseTracking,
@@ -892,88 +884,12 @@ const ContentFilterResultHeadersSchema = z
             headers["x-moderation-completion-protected-material-code-detected"],
     }));
 
-/** Salted SHA-256 hash of the full IP — irreversible without the salt. */
-async function hashIp(
-    ip: string | undefined,
-    salt: string,
-): Promise<string | undefined> {
-    if (!ip) return undefined;
-    const data = new TextEncoder().encode(`${salt}:${ip}`);
-    const hash = await crypto.subtle.digest("SHA-256", data);
-    return Array.from(new Uint8Array(hash))
-        .map((b) => b.toString(16).padStart(2, "0"))
-        .join("");
-}
-
-/** Strip `::ffff:` prefix from IPv4-mapped IPv6 addresses. */
-function stripIPv4MappedPrefix(ip: string): string {
-    const match = ip.match(/^::ffff:(\d+\.\d+\.\d+\.\d+)$/i);
-    return match ? match[1] : ip;
-}
-
-/** Truncate IP to /24 subnet (IPv4) or /48 subnet (IPv6, first 3 groups). */
-function truncateIpToSubnet(ip: string | undefined): string | undefined {
-    if (!ip) return undefined;
-    const normalized = stripIPv4MappedPrefix(ip);
-    if (normalized.includes(".")) {
-        const parts = normalized.split(".");
-        if (parts.length === 4) return `${parts[0]}.${parts[1]}.${parts[2]}.0`;
-    }
-    if (normalized.includes(":")) {
-        const full = expandIPv6(normalized);
-        const groups = full.split(":");
-        return `${groups[0]}:${groups[1]}:${groups[2]}::`;
-    }
-    return undefined;
-}
-
-/** Expand IPv6 `::` shorthand to full 8-group form. */
-function expandIPv6(ip: string): string {
-    if (!ip.includes("::")) {
-        return ip
-            .split(":")
-            .map((g) => g.padStart(4, "0"))
-            .join(":");
-    }
-    const halves = ip.split("::");
-    const left = halves[0] ? halves[0].split(":") : [];
-    const right = halves[1] ? halves[1].split(":") : [];
-    const missing = 8 - left.length - right.length;
-    const middle = Array(missing).fill("0000");
-    return [...left, ...middle, ...right]
-        .map((g) => g.padStart(4, "0"))
-        .join(":");
-}
-
 type ErrorData = {
     errorResponseCode?: string;
     errorSource?: string;
     errorMessage?: string;
     // errorStack and errorDetails removed to reduce D1 memory usage
 };
-
-async function resolveByopClientTracking(
-    db: ReturnType<typeof drizzle>,
-    byopClientKeyId: string | null | undefined,
-): Promise<Partial<UserData>> {
-    if (!byopClientKeyId) return {};
-
-    const [clientKey] = await db
-        .select({
-            id: apikeyTable.id,
-            name: apikeyTable.name,
-            userId: apikeyTable.userId,
-        })
-        .from(apikeyTable)
-        .where(eq(apikeyTable.id, byopClientKeyId))
-        .limit(1);
-
-    return {
-        apiKeyClientId: clientKey?.id ?? byopClientKeyId,
-        apiKeyCreatedForApp: clientKey?.name ?? undefined,
-        apiKeyCreatedForUserId: clientKey?.userId ?? undefined,
-    };
-}
 
 function collectErrorData(response: Response, error?: Error): ErrorData {
     if (response.ok && !error) return {};
